support double precision number for firmware

+++ FenomPlus.Mobile/FenomPlus/Views/DebugDumpView.xaml.cs
@@ -64,10 +64,12 @@ namespace FenomPlus.Views
         /// <param name="e"></param>
         public void OnSendClicked(System.Object sender, System.EventArgs e)
         {
+            
             MESSAGE message = new MESSAGE(
                 (ID_MESSAGE)Math.Abs(MessageId.SelectedIndex),
                 (ID_SUB)Math.Abs(SubId.SelectedIndex),
-                (UInt64)Math.Abs(Convert.ToInt64(Var.Value)));
+                (Byte)Math.Abs(Convert.ToByte(Var.Value)));
+            
             Services.BleHub.SendMessage(message);
         }
     }
